added Decimal32/64 and Float32 tests

+++ vertx-clickhouse-native-client/src/main/java/io/vertx/clickhouse/clickhousenative/impl/codec/columns/Float32Column.java
@@ -22,6 +22,6 @@ public class Float32Column extends ClickhouseColumn {
 
   @Override
   public Object nullValue() {
-    return (float) 0;
+    return 0.0f;
   }
 }
